NPC Art Assets
I didn't have time to make the walking animations for the NPCs so I
guess they'll just have to stand in one place for now or something.
Sorry the house is really basic, I was short on time. All character
designs are subject to change; just let me know if you have any concerns
or suggestions!