Refined Automatic Movement.
Added faceDIrection method to Player class, which lets the player face
a direction without moving in that direction.
faceDirection is only used in the Event class so far, where it is
recognized as when stop_direction == 0.
Changed Xcode project to potentially support LTBL, though there still
needs to be some more configuration done for that.