GOT SPEECH STEP WORKING!!!!
Changed it so that it only takes the changing portion of messages (as
in, the scene name).
Have only tested for one Speech Step, so far, but hopefully it works for
multiple!
Tips: use Move Step before using Speech Step to ensure the player's far
enough away from the actors to prevent accidentally initializing
dialogue by pressing space.