Commit Concerns: Successfully implemented event scripting via tmx
Can now do Speech and Direction steps in the tmx loader (instructions in
the EventScriptingviaTiled text file.
Slightly changed Speech Step (I still need to refine this one), so it's
currently kind of useless except for the Actor that it's paired with.
I'm trying to see how I can directly manipulate the dialogue going to
the textbox without passing in the textbox. However, I might just have
to pass in the textbox because passing in things by reference is not
working.
Added a getCustomDialogue event to the Event class - use this to set the
message[1] for the dialogue.