Undid Menu things and Implemented Map Moving
For now, Menu system development is going to be halted and focus is
going to be more on having multiple maps (moving between town and
field).
Currently, players can move between maps by going to a certain spot on
one map (in Tiled, it's an object in the Events object layer) and
pressing the Return button.
Scripted events in specific to each map work, and so do the dialogue and
texture drawing.
Will continue testing as the towns and battle maps get developed.