Commit Concerns: What the game is so far...
This commit contains a workable executable of a mock version of the game
(in Release_Exe zip file), with the ability to go between maps, do
priority events, and differentiate between event dialogue and regular
dialogue.
Also have an extra boolean/if condition for handling battle input at
least. Will continue working on that once code for the battle system
comes in.
Taking out some extraneous assets (mostly copies of previously created
ones) and a .wav file of Harvest Moon DS Summer (used it for sake of
testing)